requirements.txt
alembic==1.2.1
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Deprecated==1.2.6
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
future==0.18.1
github3.py==1.3.0
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
Jinja2==2.10.3
jwcrypto==0.6.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.2.1
more-itertools==7.2.0
oauthlib==3.1.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.2
pytest==5.2.1
pytest-flask==0.15.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-twitter==3.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
SQLAlchemy==1.3.10
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
